refactor sidebar layout integration
