+++ frontend/package.json
@@ -87,7 +87,8 @@
     "workbox-build": "^7.3.0",
     "workbox-window": "^7.3.0",
     "xlsx": "^0.18.5",
-    "zustand": "^5.0.7"
+    "zustand": "^5.0.7",
+    "usehooks-ts": "^3.1.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
