+++ frontend/src/components/ui/sidebar.tsx
@@ -81,7 +81,13 @@ const SidebarProvider = React.forwardRef<
     ref
   ) => {
     const isMobile = useIsMobile()
-    const [openMobile, setOpenMobile] = React.useState(false)
+    const [openMobile, setOpenMobile] = React.useState(isMobile)
+
+    React.useEffect(() => {
+      if (isMobile) {
+        setOpenMobile(true)
+      }
+    }, [isMobile])
 
     // This is the internal state of the sidebar.
     // We use openProp and setOpenProp for control from outside the component.
