+++ frontend/src/hooks/use-mobile.tsx
@@ -1,19 +1,7 @@
-import * as React from "react"
+import { useMediaQuery } from 'usehooks-ts'
 
-const MOBILE_BREAKPOINT = 768
+const MOBILE_QUERY = '(max-width: 767px)'
 
 export function useIsMobile() {
-  const [isMobile, setIsMobile] = React.useState<boolean | undefined>(undefined)
-
-  React.useEffect(() => {
-    const mql = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT - 1}px)`)
-    const onChange = () => {
-      setIsMobile(window.innerWidth < MOBILE_BREAKPOINT)
-    }
-    mql.addEventListener("change", onChange)
-    setIsMobile(window.innerWidth < MOBILE_BREAKPOINT)
-    return () => mql.removeEventListener("change", onChange)
-  }, [])
-
-  return !!isMobile
+  return useMediaQuery(MOBILE_QUERY)
 }
