+++ frontend/src/pages/NotFound.jsx
@@ -1,5 +1,6 @@
 import { useLocation } from "react-router-dom";
 import { useEffect } from "react";
+import AppLayout from '@/components/layout/AppLayout';
 
 const NotFound = () => {
   const location = useLocation();
@@ -12,6 +13,7 @@ const NotFound = () => {
   }, [location.pathname]);
 
   return (
+    <AppLayout>
     <div className="min-h-screen flex items-center justify-center bg-gray-100">
       <div className="text-center">
         <h1 className="text-4xl font-bold mb-4">404</h1>
@@ -21,6 +23,7 @@ const NotFound = () => {
         </a>
       </div>
     </div>
+    </AppLayout>
   );
 };
